test: adds unit/integration tests for configuration counterparts
