+++ lib/configuration/index.js
@@ -1,0 +1,7 @@
+const { applyConfiguration } = require('./applyConfiguration');
+const applyLoggingOptions = require('./applyLoggingOptions');
+
+module.exports = {
+  applyConfiguration,
+  applyLoggingOptions,
+};
